+++ CHANGELOG.md
@@ -10,7 +10,7 @@
     - dropdown: allow resubmitting the same value
     - plot: missing `logScaleX` and `logScaleY` properties
     - xy/multixy: add `axisLock` property; allow locking one the axis by holding `Shift` while dragging if `axisLock` is not set, bypasses it otherwise.
-    - fader: add `colorKnob` property
+    - knob/fader/range: add `colorKnob` property
 
 ## 1.25.7